backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.10.0",
    "fastapi==0.115.0",
    "openai>=1.10.0",
    "pinecone>=8.0.0",
    "psycopg2-binary==2.9.9",
    "pydantic==2.7.0",
    "python-dotenv==1.0.1",
    "sentence-transformers==2.2.2",
    "alembic>=1.13.0",
    "tenacity>=8.2.0",
    "slowapi>=0.1.9",
    "sqlalchemy==2.0.23",
    "uvicorn==0.30.1",
]
